# data/swagger.json
{
   "swagger": "2.0",
   "info": {
      "description": "API for Fixthis",
      "version": "0.1.0",
      "title": "Fixthis",
      "termsOfService": ""
   },
   "basePath": "/api",
...
   "schemes": [
      "http"
   ],
   "paths": {
...
      "/auth/resetpassword/{token}": {
         "post": {
            "tags": [
               "Authentication"
            ],
            "summary": "Verify user account",
            "description": "Api to reset user password",
            "operationId": "api.resetpassword",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "path",
                  "name": "token",
                  "type": "string",
                  "required": true,
                  "description": "Reset token"
               },
               {
                  "in": "formData",
                  "name": "password",
                  "type": "string",
                  "required": true,
                  "description": "Password to reset"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
      "/auth/logout": {
         "post": {
            "tags": [
               "Authentication"
            ],
            "summary": "Api to logout",
            "description": "Api for user logout",
            "operationId": "api.logout",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "device_token",
                  "required": true,
                  "description": "Device from which user is being logged out",
                  "type": "string"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/user/createprofile": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Create User Profile",
            "description": "Api to create user profile",
            "operationId": "api.createUserProfile",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "first_name",
                  "type": "string",
                  "required": true,
                  "description": "First name"
               },
               {
                  "in": "formData",
                  "name": "last_name",
                  "type": "string",
                  "required": true,
                  "description": "Last name"
               },
               {
                  "in": "formData",
                  "name": "subUrb",
                  "type": "string",
                  "required": true,
                  "description": "Suburban Address"
               },
               {
                  "in": "formData",
                  "name": "asFixer",
                  "type": "integer",
                  "enum": [
                     0,
                     1
                  ],
                  "required": true,
                  "description": "Do you want to create profile as fixer"
               },
               {
                  "in": "formData",
                  "name": "asPoster",
                  "type": "integer",
                  "enum": [
                     0,
                     1
                  ],
                  "required": true,
                  "description": "Do you want to create profile as poster"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
      "/user/uploadpic": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Upload profile pic",
            "description": "Api to upload user profile pic",
            "operationId": "user.uploadprofilepic",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "profilePic",
                  "type": "file",
                  "required": true,
                  "description": "Pic to upload"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/user/uploadcoverpic": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Upload cover pic",
            "description": "Api to upload user cover pic",
            "operationId": "user.uploadcoverpic",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "coverPic",
                  "type": "file",
                  "required": true,
                  "description": "Pic to upload"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/profile/addportfolio": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Upload portfolio",
            "description": "Api to upload user portfolio",
            "operationId": "user.uploadportfolio",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "portfolio",
                  "type": "file",
                  "required": true,
                  "description": "Pic to upload"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/profile/removeportfolio": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Remove Portfolio",
            "description": "Api to remove user portfolio",
            "operationId": "user.removeportfolio",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "portfolioid",
                  "type": "integer",
                  "required": true,
                  "description": "portfolio id to remove"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/profile/handlegeneralinfo": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Handle general user info in more tab",
            "description": "Api to handle general user info in more tab",
            "operationId": "user.handlegeneralinfo",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "first_name",
                  "type": "string",
                  "description": "First name"
               },
               {
                  "in": "formData",
                  "name": "last_name",
                  "type": "string",
                  "description": "last name"
               },
               {
                  "in": "formData",
                  "name": "location",
                  "type": "string",
                  "description": "location"
               },
               {
                  "in": "formData",
                  "name": "tagline",
                  "type": "string",
                  "description": "Tag Line"
               },
               {
                  "in": "formData",
                  "name": "about_me",
                  "type": "string",
                  "description": "About me"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/profile/handleprivateinfo": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Handle private user in more tab ",
            "description": "Api to handle private user info in more tab",
            "operationId": "user.privateinfo",
            "consumes": [
               "multipart/form-data"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "email",
                  "type": "string",
                  "required": true,
                  "description": "Email"
               },
               {
                  "in": "formData",
                  "name": "birth_date",
                  "type": "string",
                  "required": true,
                  "description": "Date of birth (yyyy-mm-dd)"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
      "/notifications/list": {
         "get": {
            "tags": [
               "Notifications"
            ],
            "summary": "Get user notifications",
            "description": "Api to get list of user notifications",
            "operationId": "api.getnotifications",
            "consumes": [
               "application/json"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "query",
                  "name": "pp",
                  "type": "number",
                  "description": "Notifications per page"
               },
               {
                  "in": "query",
                  "name": "pg",
                  "type": "number",
                  "description": "page number"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
      "/user/onlinestatus": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "Set user online status",
            "description": "",
            "operationId": "user.setonlinestatus",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "status",
                  "type": "string",
                  "enum": [
                     "0",
                     "1"
                  ],
                  "required": true,
                  "description": "Set current user online status"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/user/changepassword": {
         "post": {
            "tags": [
               "User"
            ],
            "summary": "API to update user password",
            "description": "Api to change user password",
            "operationId": "api.changeuserpassword",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "current_password",
                  "type": "string",
                  "required": true,
                  "description": "Current User Password"
               },
               {
                  "in": "formData",
                  "name": "new_password",
                  "type": "string",
                  "required": true,
                  "description": "New User Password"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
      "/job/list": {
         "get": {
            "tags": [
               "Jobs"
            ],
            "summary": "API to browse job list",
            "description": "Api to browse job list",
            "operationId": "api.browsejobs",
            "consumes": [],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "query",
                  "name": "myjobs",
                  "type": "boolean",
                  "required": true,
                  "description": "Do you want users jobs only"
               },
               {
                  "in": "query",
                  "name": "user_id",
                  "type": "integer",
                  "description": "Id of user for whom you want to see the jobs. Works only with myjobs = false"
               },
               {
                  "in": "query",
                  "name": "lat",
                  "type": "string",
                  "description": "User current latitude"
               },
               {
                  "in": "query",
                  "name": "long",
                  "type": "string",
                  "description": "User current longitude"
               },
               {
                  "in": "query",
                  "name": "radius",
                  "type": "string",
                  "description": "Radius (in Miles ), -1 in which radius is ignored"
               },
               {
                  "in": "query",
                  "name": "title",
                  "type": "string",
                  "description": "Search by job title"
               },
               {
                  "in": "query",
                  "name": "is_online",
                  "type": "integer",
                  "enum": [
                     0,
                     1
                  ],
                  "description": "Filter by job method"
               },
               {
                  "in": "query",
                  "name": "job_category",
                  "type": "integer",
                  "description": "Filter by job category"
               },
               {
                  "in": "query",
                  "name": "job_status",
                  "type": "string",
                  "enum": [
                     "0",
                     "1",
                     "2",
                     "3",
                     "4"
                  ],
                  "description": "Job status 0==> Draft, 1=> Post , 2==> Assigned by user , 3==>Completed, 4==>Applied"
               },
               {
                  "in": "query",
                  "name": "sort_field",
                  "type": "string",
                  "enum": [
                     "most_recent",
                     "distance",
                     "price_asc",
                     "price_desc",
                     "task_status"
                  ],
                  "description": "Field to sort by"
               },
               {
                  "in": "query",
                  "name": "pp",
                  "type": "integer",
                  "description": "records per page"
               },
               {
                  "in": "query",
                  "name": "pg",
                  "type": "integer",
                  "description": "page number"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/job/new": {
         "post": {
            "tags": [
               "Jobs"
            ],
            "summary": "Post New Job",
            "description": "Api to create new job",
            "operationId": "api.createNewJob",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "job_name",
                  "type": "string",
                  "required": true,
                  "description": "Job name"
               },
               {
                  "in": "formData",
                  "name": "job_description",
                  "type": "string",
                  "required": true,
                  "description": "Job description"
               },
               {
                  "in": "formData",
                  "name": "is_online_task",
                  "type": "integer",
                  "enum": [
                     0,
                     1
                  ],
                  "required": true,
                  "description": "Is this task to be done online"
               },
               {
                  "in": "formData",
                  "name": "category_id",
                  "type": "integer",
                  "required": true,
                  "description": "category of job"
               },
               {
                  "in": "formData",
                  "name": "location",
                  "type": "string",
                  "required": true,
                  "description": "Job location"
               },
               {
                  "in": "formData",
                  "name": "latitude",
                  "type": "string",
                  "required": true,
                  "description": "Job latitude"
               },
               {
                  "in": "formData",
                  "name": "longitude",
                  "type": "string",
                  "required": true,
                  "description": "Job longitude"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/job/update": {
         "post": {
            "tags": [
               "Jobs"
            ],
            "summary": "Update Job",
            "description": "Api to update job",
            "operationId": "api.updateJob",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "job_id",
                  "type": "number",
                  "required": true,
                  "description": "Job id to update"
               },
               {
                  "in": "formData",
                  "name": "job_name",
                  "type": "string",
                  "description": "Job name"
               },
               {
                  "in": "formData",
                  "name": "job_description",
                  "type": "string",
                  "description": "Job description"
               },
               {
                  "in": "formData",
                  "name": "is_online_task",
                  "type": "integer",
                  "enum": [
                     0,
                     1
                  ],
                  "description": "Is this task to be done online"
               },
               {
                  "in": "formData",
                  "name": "job_status",
                  "type": "integer",
                  "enum": [
                     0,
                     1,
                     2,
                     3
                  ],
                  "description": "Job status 0==> Draft, 1=> Post , 2==> Assigned by user , 3==>Completed"
               },
               {
                  "in": "formData",
                  "name": "hourly_rate",
                  "type": "number",
                  "description": "Hourly rate of job"
               },
               {
                  "in": "formData",
                  "name": "hours",
                  "type": "number",
                  "description": "Hours"
               },
               {
                  "in": "formData",
                  "name": "task_price",
                  "type": "number",
                  "description": "Task budget"
               },
               {
                  "in": "formData",
                  "name": "task_end_day",
                  "type": "string",
                  "description": "Task end day == > dd-mm-yyyy"
               },
               {
                  "in": "formData",
                  "name": "task_end_time",
                  "type": "string",
                  "description": "Task end time == > 12:00am"
               },
               {
                  "in": "formData",
                  "name": "no_of_fixers",
                  "type": "number",
                  "description": "Number of fixers"
               },
               {
                  "in": "formData",
                  "name": "abn_number",
                  "type": "string",
                  "description": "Poster abn number"
               },
               {
                  "in": "formData",
                  "name": "location",
                  "type": "string",
                  "description": "Job location"
               },
               {
                  "in": "formData",
                  "name": "latitude",
                  "type": "string",
                  "description": "Job latitude"
               },
               {
                  "in": "formData",
                  "name": "longitude",
                  "type": "string",
                  "description": "Job longitude"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
      "/job/delete": {
         "post": {
            "tags": [
               "Jobs"
            ],
            "summary": "Delete Job",
            "description": "Api to delete",
            "operationId": "api.deleteJob",
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "job_id",
                  "type": "integer",
                  "required": true,
                  "description": "Job id"
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation"
               }
            }
         }
      },
...
   },
...
   "securityDefinitions": {
      "bearer": {
         "in": "header",
         "name": "Authorization",
         "type": "apiKey"
      }
   }
}